requirements.txt
# Core
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Agent
langgraph>=0.0.40
ollama>=0.3.0

# Screen capture & OCR
mss>=9.0.0
opencv-python>=4.8.0
pillow>=10.0.0
pytesseract>=0.3.10

# Window detection
pygetwindow>=0.0.9

# Dev tools
black>=23.7.0
